Dedicated DockPartial & fixed post-login route
